refactor: use HOC for auth guard
