+++ src/app/(authentication)/onboarding/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Onboarding } from "./Onboarding";
 
 export const metadata: Metadata = {
-  title: "Bang Soal",
+  title: "BangSoal",
   description: "Onboarding",
 };
