+++ src/app/(authentication)/signup/constants.ts
@@ -8,12 +8,12 @@ export const SIGNUP_COPYWRITING = {
   login: {
     header: "Get back in the game!",
     caption:
-      "Sign in with your Bang Soal account to start practicing for your UTBK!",
+      "Sign in with your BangSoal account to start practicing for your UTBK!",
     anchor: "A new user? Make an account.",
   },
   register: {
     header: "Get in the game!",
-    caption: "Create an account to start using Bang Soal today! ",
+    caption: "Create an account to start using BangSoal today! ",
     anchor: "Already a user? Sign in.",
   },
 };
