+++ src/app/(authentication)/signup/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { AuthenticationContainer } from "./AuthenticationContainer";
 
 export const metadata: Metadata = {
-  title: "Bang Soal",
+  title: "BangSoal",
   description: "Authentication",
 };
